+++ src/main/java/edu/boisestate/cs/graph/InvConstraintConcreteValue.java
@@ -15,15 +15,15 @@ import edu.boisestate.cs.solvers.Solver_Inverse;
  * @author marli
  *
  */
-public class InvConstraintConcreteValue_r3<T extends A_Model_Inverse<T>>  extends A_Inv_Constraint<T> {
+public class InvConstraintConcreteValue<T extends A_Model_Inverse<T>>  extends A_Inv_Constraint<T> {
 
 	
 	//private SolutionSet<T> solutionSet;
 	
-	private int solutionIndex = -1;
+	//private int solutionIndex = -1;
 	
 	
-	public InvConstraintConcreteValue_r3 (int ID, Solver_Inverse<T> solver) {
+	public InvConstraintConcreteValue (int ID, Solver_Inverse<T> solver) {
 		
 		// Store reference to solver
 		this.solver = solver;
@@ -36,7 +36,7 @@ public class InvConstraintConcreteValue_r3<T extends A_Model_Inverse<T>>  extend
 	}
 	
 	
-	public InvConstraintConcreteValue_r3 (int ID, Solver_Inverse<T> solver, List<Integer> args, SolutionSet<T> solutionSet) {
+	public InvConstraintConcreteValue (int ID, Solver_Inverse<T> solver, List<Integer> args, SolutionSet<T> solutionSet) {
 		
 		// Store reference to solver
 		this.solver = solver;
@@ -47,7 +47,7 @@ public class InvConstraintConcreteValue_r3<T extends A_Model_Inverse<T>>  extend
 		//this.solutionSet = solutionSet;
 	}
 	
-	public InvConstraintConcreteValue_r3 (int ID, Solver_Inverse<T> solver, List<Integer> args, SolutionSet<T> solutionSet, int base, int input) {
+	public InvConstraintConcreteValue (int ID, Solver_Inverse<T> solver, List<Integer> args, SolutionSet<T> solutionSet, int base, int input) {
 		
 		// Store reference to solver
 		this.solver = solver;
