+++ src/main/java/edu/boisestate/cs/graph/InvConstraintReplaceCharChar.java
@@ -13,35 +13,34 @@ import edu.boisestate.cs.solvers.*;
  * @author marli
  *
  */
-public class InvConstraintSubStringStartEnd_r3<T extends A_Model_Inverse<T>> extends A_Inv_Constraint<T> {
+public class InvConstraintReplaceCharChar<T extends A_Model_Inverse<T>> extends A_Inv_Constraint<T> {
 	
-
-	private int start,end;
+	private int find,replace;
 	
-	public InvConstraintSubStringStartEnd_r3 (int ID, Solver_Inverse<T> solver, List<Integer> args) {
+	public InvConstraintReplaceCharChar (int ID, Solver_Inverse<T> solver, List<Integer> args) {
 		
 		// Store reference to solver
 		this.solver = solver;
 		this.ID = ID;
 		this.argList = args;
-		this.op = Operation.SUBSTR_STRT_END;
+		this.op = Operation.REPLACE_CHAR_CHAR;
 		this.outputSet = new HashMap<Integer,T>();
 		this.solutionSet = new SolutionSetInternal<T>(ID);
-		this.argString = "0:START 1:END";
-		this.start = argList.get(0);
-		this.end = argList.get(1);
+		this.argString = "0:FIND 1:REPLACE";
+		this.find = argList.get(0);
+		this.replace = argList.get(1);
 	}
 	
-	public InvConstraintSubStringStartEnd_r3 (int ID, Solver_Inverse<T> solver, List<Integer> args, int base, int input) {
+	public InvConstraintReplaceCharChar (int ID, Solver_Inverse<T> solver, List<Integer> args, int base, int input) {
 		
 		// Store reference to solver
 		this.solver = solver;
 		this.ID = ID;
 		this.argList = args;
-		this.op = Operation.SUBSTR_STRT_END;
-		this.argString = "0:START 1:END";
-		this.start = argList.get(0);
-		this.end = argList.get(1);
+		this.op = Operation.REPLACE_CHAR_CHAR;
+		this.argString = "0:FIND 1:REPLACE";
+		this.find = argList.get(0);
+		this.replace = argList.get(1);
 		this.nextID = base;
 		this.prevID = input;
 	}
@@ -50,12 +49,13 @@ public class InvConstraintSubStringStartEnd_r3<T extends A_Model_Inverse<T>> ext
 	@Override
 	public boolean evaluate(I_Inv_Constraint<T> inputConstraint, int sourceIndex) {
 		
-		System.out.format("EVALUATE SUBSTRING %d ...\n",ID);
+		// solver.inv_replaceCharKnown(ID, prevConstraint.getID(), (char) find, (char) replace);
+		System.out.format("EVALUATE REPLACE CHAR %d ...\n",ID);
 		
 		T inputModel = inputConstraint.output(sourceIndex);
 
 		// perform inverse function on output from the input constraint at given index
-		T resModel = solver.inv_substring(inputModel, start, end);
+		T resModel = solver.inv_replaceCharKnown(inputModel, (char) find, (char) replace);
 
 		// intersect result with forward analysis results from previous constraint
 		resModel = solver.intersect(resModel, nextConstraint.getID());
@@ -73,17 +73,19 @@ public class InvConstraintSubStringStartEnd_r3<T extends A_Model_Inverse<T>> ext
 				// we have values, so continue solving ...
 				return nextConstraint.evaluate(this, 1);
 			} else {
-				System.out.println("SUBSTRING SOLUTION SET INCONSISTENT...");
+				System.out.println("REPLACE CHAR SOLUTION SET INCONSISTENT...");
 				solutionSet.remSolution(inputConstraint.getID());
 				return false;
 			}
 			
 		} else {
-			System.out.println("SUBSTRING RESULT MODEL EMPTY...");
+			System.out.println("REPLACE CHAR RESULT MODEL EMPTY...");
 			// halt solving, fallback
 			return false;
 		}
 		
 	}
+
+
 
 }
