+++ src/main/java/edu/boisestate/cs/graph/InvConstraintToLowerCase.java
@@ -13,34 +13,40 @@ import edu.boisestate.cs.solvers.*;
  * @author marli
  *
  */
-public class InvConstraintReplaceCharChar_r3<T extends A_Model_Inverse<T>> extends A_Inv_Constraint<T> {
+public class InvConstraintToLowerCase<T extends A_Model_Inverse<T>> extends A_Inv_Constraint<T> {
 	
-	private int find,replace;
 	
-	public InvConstraintReplaceCharChar_r3 (int ID, Solver_Inverse<T> solver, List<Integer> args) {
+	public InvConstraintToLowerCase (int ID, Solver_Inverse<T> solver) {
 		
 		// Store reference to solver
 		this.solver = solver;
 		this.ID = ID;
-		this.argList = args;
-		this.op = Operation.REPLACE_CHAR_CHAR;
+		this.op  = Operation.TOLOWERCASE;
 		this.outputSet = new HashMap<Integer,T>();
 		this.solutionSet = new SolutionSetInternal<T>(ID);
-		this.argString = "0:FIND 1:REPLACE";
-		this.find = argList.get(0);
-		this.replace = argList.get(1);
+		this.argString = "[NONE]";
 	}
 	
-	public InvConstraintReplaceCharChar_r3 (int ID, Solver_Inverse<T> solver, List<Integer> args, int base, int input) {
+	
+	
+	public InvConstraintToLowerCase (int ID, Solver_Inverse<T> solver, List<Integer> args) {
 		
 		// Store reference to solver
 		this.solver = solver;
 		this.ID = ID;
 		this.argList = args;
-		this.op = Operation.REPLACE_CHAR_CHAR;
-		this.argString = "0:FIND 1:REPLACE";
-		this.find = argList.get(0);
-		this.replace = argList.get(1);
+		this.op  = Operation.TOLOWERCASE;
+		this.argString = "[NONE]";
+	}
+	
+	public InvConstraintToLowerCase (int ID, Solver_Inverse<T> solver, List<Integer> args, int base, int input) {
+		
+		// Store reference to solver
+		this.solver = solver;
+		this.ID = ID;
+		this.argList = args;
+		this.op  = Operation.TOLOWERCASE;
+		this.argString = "[NONE]";
 		this.nextID = base;
 		this.prevID = input;
 	}
@@ -48,14 +54,13 @@ public class InvConstraintReplaceCharChar_r3<T extends A_Model_Inverse<T>> exten
 	
 	@Override
 	public boolean evaluate(I_Inv_Constraint<T> inputConstraint, int sourceIndex) {
-		
-		// solver.inv_replaceCharKnown(ID, prevConstraint.getID(), (char) find, (char) replace);
-		System.out.format("EVALUATE REPLACE CHAR %d ...\n",ID);
+
+		System.out.format("EVALUATE TOLOWER %d ...\n",ID);
 		
 		T inputModel = inputConstraint.output(sourceIndex);
 
 		// perform inverse function on output from the input constraint at given index
-		T resModel = solver.inv_replaceCharKnown(inputModel, (char) find, (char) replace);
+		T resModel = solver.inv_toLowerCase(inputModel);
 
 		// intersect result with forward analysis results from previous constraint
 		resModel = solver.intersect(resModel, nextConstraint.getID());
@@ -73,19 +78,23 @@ public class InvConstraintReplaceCharChar_r3<T extends A_Model_Inverse<T>> exten
 				// we have values, so continue solving ...
 				return nextConstraint.evaluate(this, 1);
 			} else {
-				System.out.println("REPLACE CHAR SOLUTION SET INCONSISTENT...");
+				System.out.println("TOLOWER SOLUTION SET INCONSISTENT...");
 				solutionSet.remSolution(inputConstraint.getID());
 				return false;
 			}
 			
 		} else {
-			System.out.println("REPLACE CHAR RESULT MODEL EMPTY...");
+			System.out.println("TOLOWER RESULT MODEL EMPTY...");
 			// halt solving, fallback
 			return false;
 		}
-		
-	}
 
+
+
+
+
+
+	}
 
 
 }
