+++ src/main/java/edu/boisestate/cs/graph/InvConstraintToUpperCase.java
@@ -13,39 +13,39 @@ import edu.boisestate.cs.solvers.*;
  * @author marli
  *
  */
-public class InvConstraintToLowerCase_r3<T extends A_Model_Inverse<T>> extends A_Inv_Constraint<T> {
+public class InvConstraintToUpperCase<T extends A_Model_Inverse<T>> extends A_Inv_Constraint<T> {
 	
 	
-	public InvConstraintToLowerCase_r3 (int ID, Solver_Inverse<T> solver) {
+	public InvConstraintToUpperCase (int ID, Solver_Inverse<T> solver) {
 		
 		// Store reference to solver
 		this.solver = solver;
 		this.ID = ID;
-		this.op  = Operation.TOLOWERCASE;
+		this.op  = Operation.TOUPPERCASE;
 		this.outputSet = new HashMap<Integer,T>();
 		this.solutionSet = new SolutionSetInternal<T>(ID);
 		this.argString = "[NONE]";
 	}
 	
+
 	
-	
-	public InvConstraintToLowerCase_r3 (int ID, Solver_Inverse<T> solver, List<Integer> args) {
+	public InvConstraintToUpperCase (int ID, Solver_Inverse<T> solver, List<Integer> args) {
 		
 		// Store reference to solver
 		this.solver = solver;
 		this.ID = ID;
 		this.argList = args;
-		this.op  = Operation.TOLOWERCASE;
+		this.op = Operation.TOUPPERCASE;
 		this.argString = "[NONE]";
 	}
 	
-	public InvConstraintToLowerCase_r3 (int ID, Solver_Inverse<T> solver, List<Integer> args, int base, int input) {
+	public InvConstraintToUpperCase (int ID, Solver_Inverse<T> solver, List<Integer> args, int base, int input) {
 		
 		// Store reference to solver
 		this.solver = solver;
 		this.ID = ID;
 		this.argList = args;
-		this.op  = Operation.TOLOWERCASE;
+		this.op = Operation.TOUPPERCASE;
 		this.argString = "[NONE]";
 		this.nextID = base;
 		this.prevID = input;
@@ -55,12 +55,12 @@ public class InvConstraintToLowerCase_r3<T extends A_Model_Inverse<T>> extends A
 	@Override
 	public boolean evaluate(I_Inv_Constraint<T> inputConstraint, int sourceIndex) {
 
-		System.out.format("EVALUATE TOLOWER %d ...\n",ID);
+		System.out.format("EVALUATE TOUPPER %d ...\n",ID);
 		
 		T inputModel = inputConstraint.output(sourceIndex);
 
 		// perform inverse function on output from the input constraint at given index
-		T resModel = solver.inv_toLowerCase(inputModel);
+		T resModel = solver.inv_toUpperCase(inputModel);
 
 		// intersect result with forward analysis results from previous constraint
 		resModel = solver.intersect(resModel, nextConstraint.getID());
@@ -73,28 +73,21 @@ public class InvConstraintToLowerCase_r3<T extends A_Model_Inverse<T>> extends A
 	
 				// store result in this constraints output set at index 1
 				outputSet.put(1, resModel);	
-	
-	
+		
 				// we have values, so continue solving ...
 				return nextConstraint.evaluate(this, 1);
 			} else {
-				System.out.println("TOLOWER SOLUTION SET INCONSISTENT...");
+				System.out.println("TOUPPER SOLUTION SET INCONSISTENT...");
 				solutionSet.remSolution(inputConstraint.getID());
 				return false;
 			}
 			
 		} else {
-			System.out.println("TOLOWER RESULT MODEL EMPTY...");
+			System.out.println("TOUPPER RESULT MODEL EMPTY...");
 			// halt solving, fallback
 			return false;
 		}
 
-
-
-
-
-
 	}
-
 
 }
